+++ packages/pesapal/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @openfn/language-pesapal
 
+## 1.1.2 - 10 July 2025
+
+### Patch Changes
+
+- Updated dependencies \[cf9c09f]
+  - @openfn/language-common@3.0.1
+
+## 1.1.1 - 10 July 2025
+
+### Patch Changes
+
+- f182cfc: Add required fields in configuration-schema
+
+## 1.1.0 - 10 July 2025
+
+### Minor Changes
+
+- 19f2d7e: Exported `as()` function from common.
+
+### Patch Changes
+
+- Updated dependencies \[ea85695]
+- Updated dependencies \[3fce58f]
+- Updated dependencies \[19f2d7e]
+- Updated dependencies \[f26bd2b]
+- Updated dependencies \[19f2d7e]
+  - @openfn/language-common@3.0.0
+
 ## 1.0.3 - 20 June 2025
 
 ### Patch Changes